Create types file in utils and refactor
