+++ src/components/Header/Header.tsx
@@ -1,13 +1,11 @@
-import React from "react";
-import bee from "../../images/bee.png";
-import "./Header.css";
+import React from 'react';
+import bee from '../../images/bee.png';
+import './Header.css';
 
-interface HeaderProps {}
-
-export const Header: React.FC<HeaderProps> = () => {
+export const Header: React.FC = () => {
   return (
     <header>
-      <img src={bee} alt="round smiling bee" className="logo" />
+      <img src={bee} alt='round smiling bee' className='logo' />
       <h1>HiveMind</h1>
     </header>
   );
